Review Comments: Moved the connect function to different file. Added a new spec and reduced the dependency of having SingleEmissionSubject for compiling the filterSourceBy function
